32-jenkins-remoting-project/requirements.txt
jenkins==1.0.0
pyyaml==6.0.1
paramiko==3.4.0
cryptography==42.0.5
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.8
flask==3.0.2
python-jenkins==1.8.0
threading
concurrent-log-handler